Migrate vaccination infraestructure. (#261)
# Our World In Data
The motivation behind this work is to let Our World In Data have more information for boosters and unique doses, separately. Also to users who have relied on this data. 

Due to the sheer amount of work required to extract data from Ministerio Salud Publica's [Vacunómetro](https://app.powerbi.com/view?r=eyJrIjoiMjgxODQyZDItMTZiYi00ZjhmLWI0MzEtYWJlNzAxZDcwNWIzIiwidCI6IjcwNjIyMGRiLTliMjktNGU5MS1hODI1LTI1NmIwNmQyNjlmMyJ9&pageName=ReportSection5e050ac003d0b042a320) and the difficulties reverse engineering their data (including them constantly change their UI with breaking changes of the data) the data provided here have had small precomputations in it (such as providing the field for second doses `segunda_dosis` as the sum of the second doses and unique doses) in the interest of not breaking any consumers as well.

This PR belongs to the [migration plan](https://github.com/andrab/ecuacovid/issues/203#issuecomment-962740011) mentioned with a few things to note (and will be merged as soon as @edomt approves since changes in the OWD batch program are required).

[Based on](https://github.com/andrab/ecuacovid/issues/203#issuecomment-963044138) this would be the new changes:

* `fabricantes.csv` columns `dosis_unica` (monodose) and `dosis_refuerzo` (booster) added per manufacturer. 
* `vacunas.csv` column `dosis_unica` added.

**important** to note that the column `segunda_dosis` no longer contains the sum of second doses and monodoses, just second doses this time. The numbers were adjusted (removing the monodose quantity from the `segunda_dosis` columns in all the data, including the data already extracted)

# General

After discussions with @Panchurtado-gr, the following new data is also provided:

* Provinces now also have monodose and booster data. 
* Cities now also have monodose and booster data.
* Tabulated data for above and percentages of them [are also provided](https://github.com/andrab/ecuacovid/tree/master/datos_crudos/vacunometro/por_fecha). When this PR is merged we've have more of these for boosters and monodoses for cities and provinces.

Finally we are removing an alternate file (planvacunarse.csv) that originally existed and was contributed by @sbibauw when vacunometro was launched as we tried our best to figure out how to get data.